beam: widen history cursor to uint64_t and guard index overflow
Address Copilot review on #183 (all three comments, one root cause):

- Make osh_beam_runtime::primaries_generated a uint64_t to match the rest of
  the history-index plumbing (pool->prim_idx, hist_index). As size_t it could
  truncate past 2^32 on 32-bit builds and reuse prim_idx / RNG seeds.
- Guard osh_beam_runtime_fill_pool_at() against a base whose range
  [base, base + n) would wrap uint64_t: return OSH_EINVAL early and leave the
  pool untouched instead of silently reusing indices/streams. Unreachable in
  practice (~2^64 histories) but the seam is the parallel-fill contract, so it
  fails loudly. The serial wrapper inherits this: it advances the cursor only
  on OSH_OK, so a wrapping base is rejected before the cursor moves.
- Add test_overflow_guard: a wrapping base is rejected with the pool left
  empty; a large non-wrapping base is accepted and assigns those global indices.

+++ src/beam/runtime/osh_beam_runtime.h
@@ -47,7 +47,7 @@ extern "C" {
 struct osh_beam_runtime {
     struct beam_workspace const *workspace; /* cold storage — not owned */
     struct particle primary;                /* resolved species for beam primaries */
-    size_t primaries_generated;             /* serial-path cursor; see header doc */
+    uint64_t primaries_generated;           /* serial-path cursor; see header doc */
 
     union {
         /**
